+++ R/function_PlotBasinSummary.R
@@ -1,7 +1,7 @@
 #' Plot a summary of model results for a single sub-basin
 #' 
 #' Plot a standard suite of plots summarizing properties of a sub-basin including upstream area and model performance 
-#' for discharge and concentrations of nutrients, sediment, and tracers.
+#' for discharge and concentrations of nutrients, organic carbon, sediment, and tracers.
 #' 
 #' @param x Data frame, with column-wise daily time series of HYPE variables. Date-times in 
 #' \code{\link{POSIXct}} format in first column. Typically an imported basin output file from HYPE using \code{\link{ReadBasinOutput}}. 
@@ -44,7 +44,7 @@
 #' @details
 #' \code{PlotBasinSummary} plots a multi-panel plot with a number of plots to evaluate model properties and performances for a 
 #' chosen sub-basin. Performance plots include discharge, HYPE-modeled nutrient species for nitrogen (total, inorganic, organic) 
-#' and phosphorus (total, particulate, soluble), and HYPE modeled suspended and total sediment concentrations.
+#' phosphorus (total, particulate, soluble), organic carbon, and suspended and total sediment concentrations.
 #' 
 #' Plotted panels show: 
 #' \itemize{
@@ -54,7 +54,7 @@
 #' \item{\emph{Goodness-of-fit measures for discharge and concentrations}: KGE (Kling-Gupta Efficiency), NSE (Nash-Sutcliffe Efficiency), PBIAS 
 #' (Percentage Bias, aka relative error), MAE (Mean Absolute Error), r (Pearson product-moment correlation coefficient), VE (Volumetric Efficiency).}
 #' \item{\emph{Simulation-observation relationships for discharge and concentrations}: Simulated and observed concentration-discharge relationships, 
-#' relationship between observed and simulated nutrient, sediment, and tracer concentrations.}
+#' relationship between observed and simulated nutrient, organic carbon, sediment, and tracer concentrations.}
 #' \item{\emph{Duration curves for flow and concentrations}: Pairwise simulated and observed curves.}
 #' \item{\emph{Annual regimes for flow and concentrations}: Pairwise simulated and observed regime plots at monthly aggregation, with 
 #' number of observations for concentration regimes.}
@@ -73,7 +73,7 @@
 #' For a detailed description of the variables, see the 
 #' \href{http://hype.smhi.net//wiki/doku.php?id=start:hype_file_reference:info.txt:variables}{HYPE online documentation}.
 #' 
-#' \code{basinoutput variable cout rout ccin rein ccon reon cctn retn ccsp resp ccpp repp cctp retp ctnl ctpl ccss ress ccts rets cct1 ret1}
+#' \code{basinoutput variable cout rout ccin rein ccon reon cctn retn ccsp resp ccpp repp cctp retp ctnl ctpl ccss ress ccts rets cct1 ret1 ccoc reoc}
 #' 
 #' #' \emph{Device dimensions} are hard-coded to a width of 13 inches and height depending on the number of plotted time series. When plotting 
 #' to a screen device, a maximum height of 10 inches is enforced in order to prevent automatic resizing with slow redrawing. 
@@ -118,7 +118,7 @@ PlotBasinSummary <- function(x, filename, driver = c("default", "pdf", "png", "s
   # Assign NULL value to all potential data variables extracted from 'x' further below
   # This to get rid of 'no visible binding for global variable'-notes with 'Check Package'
   # See Option Two here: https://www.r-bloggers.com/2019/08/no-visible-binding-for-global-variable/
-  date <- cout <- rout <- ccin <- rein <- ccon <- reon <- cctn <- retn <- ccsp <- resp <- 
+  date <- cout <- rout <- ccin <- rein <- ccon <- reon <- cctn <- retn <- ccoc <- reoc <- ccsp <- resp <- 
     ccpp <- repp <- cctp <- retp <- ctnl <- ctpl <- ccss <- ress <- ccts <- rets <- cct1 <- ret1 <- NULL
   
   # check and choose device driver
@@ -213,7 +213,7 @@ PlotBasinSummary <- function(x, filename, driver = c("default", "pdf", "png", "s
   
   # create vector over all target output variable names which are potentially used in the plot panels
   nm.t <- c("date", "cout", "rout", "ccin", "rein", "ccon", "reon", "cctn", "retn", "ccsp", "resp", "ccpp", 
-            "repp", "cctp", "retp", "ctnl", "ctpl", "ccss", "ress", "ccts", "rets", "cct1", "ret1")
+            "repp", "cctp", "retp", "ctnl", "ctpl", "ccss", "ress", "ccts", "rets", "cct1", "ret1", "ccoc", "reoc")
   # initialise logical vector to indicate existence of target variables
   exi.t <- logical(length = length(nm.t))
   names(exi.t) <- nm.t
@@ -237,7 +237,7 @@ PlotBasinSummary <- function(x, filename, driver = c("default", "pdf", "png", "s
       exi.t[!(nm.t %in% nm.hydro)] <- FALSE
     } else if (hype.vars[1] == "wq") {
       nm.wq <- c("date", "ccin", "rein", "ccon", "reon", "cctn", "retn", "ccsp", "resp", "ccpp", "repp", 
-                    "cctp", "retp", "ctnl", "ctpl", "cout", "ccss", "ress", "ccts", "rets", "cct1", "ret1")
+                    "cctp", "retp", "ctnl", "ctpl", "cout", "ccss", "ress", "ccts", "rets", "cct1", "ret1", "ccoc", "reoc")
       exi.t[!(nm.t %in% nm.wq)] <- FALSE
     } else if (is.character(hype.vars)) {
       nm.manu <- c("date", tolower(hype.vars))
@@ -1839,7 +1839,178 @@ PlotBasinSummary <- function(x, filename, driver = c("default", "pdf", "png", "s
                                      ylab = expression(paste("SS conc. (mg l"^"-1", ")")))')
       }
     }
+
+
+
     
+    #--------------------------------------------------------------------------------------------------------------------------------------
+    # OC row: 5 panels with GOFs, Conc-Q relationships, sim-obs, FDC, and regime for OC, conditional on if OC is requested/available
+    #--------------------------------------------------------------------------------------------------------------------------------------
+    
+    if (exi.t["reoc"] || exi.t["ccoc"]) {
+      
+      # fill layout matrix with panel IDs
+      lay.mat <- rbind(lay.mat, c(seq(max(lay.mat) + 1, by = 1, length.out = 5), max(lay.mat) + 5)) 
+      # add layout height for this row
+      lay.heights <- c(lay.heights, 2)
+      
+      
+      ## panel 1: compute and plot GoFs for OC, if variables are available
+      # empty frame first, then GoFs as legend
+      cp <- cp + 1
+      list.plotexpr[[cp]] <- parse(text = 'par(mar = rep(0, 4))')
+      cp <- cp + 1
+      list.plotexpr[[cp]] <- parse(text = 'frame()')
+      
+      if (exi.t["reoc"] && exi.t["ccoc"]){
+        gof.oc <- tryCatch(gof(sim = get("ccoc"), obs = get("reoc"), na.rm = TRUE)[c("KGE", "NSE", "PBIAS %", "MAE", "r", "VE"), ], 
+                           error = function(e){te <- rep(NA, 6); names(te) <- c("KGE", "NSE", "PBIAS %", "MAE", "r", "VE"); te})
+        cp <- cp + 1
+        list.plotexpr[[cp]] <- parse(text = 'legend(x = 0, y = 0.9, 
+                                     legend = c(paste(names(gof.oc), gof.oc, sep = ": "),"",
+                                     paste0("(", length(na.omit(reoc)), " obs.)")), bty = "n", 
+                                     title = "OC, goodness\nof fit", cex = 1)')
+      }
+      
+      
+      ## panel 2: Conc-Q plots, depending on variable availability
+      if (exi.t["cout"] && (exi.t["reoc"] || exi.t["ccoc"])) {
+        if (exi.t["reoc"] && exi.t["ccoc"]) {
+          
+          # calculate y axis limits
+          lim.oc <- range(c(reoc, ccoc), na.rm = TRUE)
+          # change lower limit to >0 if log-scale
+          if (log && lim.oc[1] <= 0) {
+            lim.oc[1] <- min(c(reoc, ccoc)[c(reoc, ccoc) > 0], na.rm = TRUE)
+            # treat case where there are no non-0 values, and Inf values are created
+            if (any(is.infinite(lim.oc))) {
+              lim.oc <- rep(0, 2)
+            }
+          }
+          
+          
+          cp <- cp + 1
+          list.plotexpr[[cp]] <- parse(text = 'par(mar = c(3.1, 3.1, .5, .5), tcl = -0.2, mgp = c(1.8, 0.3, 0))')
+          cp <- cp + 1
+          list.plotexpr[[cp]] <- parse(text = 'plot(cout, ccoc, col = "#FF00003C", pch = 16, log = log.cq, cex = .7, xlim = lim.q, ylim = lim.oc, xlab = expression(paste("simulated Q (m"^3,"s"^"-1", ")")), ylab = expression(paste("OC conc. (mg l"^"-1", ")")))')
+          cp <- cp + 1
+          list.plotexpr[[cp]] <- parse(text = 'points(cout, reoc, col = "#00000080", pch = 16)')
+          cp <- cp + 1
+          list.plotexpr[[cp]] <- parse(text = 'legend("topright", legend = c("sim.", "obs."), pch = 16, col = c("#FF000080", "#00000080"), bty = "n")')
+          
+        }  else if (exi.t["reoc"]) {
+          
+          # calculate y axis limits
+          lim.oc <- range(reoc, na.rm = TRUE)
+          # change lower limit to >0 if log-scale
+          if (log && lim.oc[1] <= 0) {
+            lim.oc[1] <- min(reoc[reoc > 0], na.rm = TRUE)
+            # treat case where there are no non-0 values, and Inf values are created
+            if (any(is.infinite(lim.oc))) {
+              lim.oc <- rep(0, 2)
+            }
+          }
+          
+          cp <- cp + 1
+          list.plotexpr[[cp]] <- parse(text = 'par(mar = c(3.1, 3.1, .5, .5), tcl = -0.2, mgp = c(1.8, 0.3, 0))')
+          cp <- cp + 1
+          list.plotexpr[[cp]] <- parse(text = 'plot(cout, reoc, col = "#00000080", pch = 16, log = log.cq, xlim = lim.q, ylim = lim.oc, xlab = expression(paste("simulated Q (m"^3,"s"^"-1", ")")), ylab = expression(paste("OC conc. (mg l"^"-1", ")")))')
+          cp <- cp + 1
+          list.plotexpr[[cp]] <- parse(text = 'legend("topright", legend = c("sim.", "obs."), pch = 16, col = c("#FF000080", "#00000080"), bty = "n")')
+          
+        } else if (exi.t["ccoc"]) {
+          
+          # calculate y axis limits
+          lim.oc <- range(ccoc, na.rm = TRUE)
+          # change lower limit to >0 if log-scale
+          if (log && lim.oc[1] <= 0) {
+            lim.oc[1] <- min(ccoc[ccoc > 0], na.rm = TRUE)
+            # treat case where there are no non-0 values, and Inf values are created
+            if (any(is.infinite(lim.oc))) {
+              lim.oc <- rep(0, 2)
+            }
+          }
+          
+          cp <- cp + 1
+          list.plotexpr[[cp]] <- parse(text = 'par(mar = c(3.1, 3.1, .5, .5), tcl = -0.2, mgp = c(1.8, 0.3, 0))')
+          cp <- cp + 1
+          list.plotexpr[[cp]] <- parse(text = 'plot(cout, ccoc, col = "#FF00003C", pch = 16, log = log.cq, cex = .7, 
+                                       xlim = lim.q, ylim = lim.oc, xlab = expression(paste("simulated Q (m"^3,"s"^"-1", ")")), 
+                                       ylab = expression(paste("OC conc. (mg l"^"-1", ")")))')
+          cp <- cp + 1
+          list.plotexpr[[cp]] <- parse(text = 'legend("topright", legend = c("sim.", "obs."), pch = 16, 
+                                       col = c("#FF000080", "#00000080"), bty = "n")')
+          
+        }
+      } else {
+        cp <- cp + 1
+        list.plotexpr[[cp]] <- parse(text = 'frame()')
+      }
+      
+      
+      ## panel 3: sim-obs comparison dotty plot for OC
+      if (exi.t["ccoc"] && exi.t["reoc"]){
+        cp <- cp + 1
+        list.plotexpr[[cp]] <- parse(text = 'par(mar = c(3.1, 3.1, .5, .5), tcl = -0.2, mgp = c(1.8, 0.3, 0))')
+        cp <- cp + 1
+        list.plotexpr[[cp]] <- parse(text = 'plot(reoc, ccoc, col = "#7FAD8EE6", pch = 16, log = log.cq, 
+                                     xlab = expression(paste("observed OC (mg l"^"-1", ")")), 
+                                     ylab = expression(paste("simulated OC (mg l"^"-1", ")")))')
+        cp <- cp + 1
+        list.plotexpr[[cp]] <- parse(text = 'abline(a = 0, b = 1, col = "#00000080")')
+      } else {
+        cp <- cp + 1
+        list.plotexpr[[cp]] <- parse(text = 'frame()')
+      }
+      
+      
+      
+      ## panel 4: CDC for OC
+      if (exi.t["reoc"] && exi.t["ccoc"]){
+        cp <- cp + 1
+        list.plotexpr[[cp]] <- parse(text = '.PlotDurationCurve(ExtractFreq(data = data.frame(reoc, ccoc)), 
+                                     xscale = xscale, yscale = yscale, add.legend = TRUE, l.legend = c("obs. OC", "sim. OC"), 
+                                     col = c("black", "red"), mar = c(3.1, 3.1, .5, .5), 
+                                     xlab = "Concentration exceedance percentile", 
+                                     ylab = expression(paste("OC conc. (mg l"^"-1", ")")))')
+      } else if (exi.t["reoc"]) {
+        cp <- cp + 1
+        list.plotexpr[[cp]] <- parse(text = '.PlotDurationCurve(ExtractFreq(data = reoc), xscale = xscale, yscale = yscale, 
+                                     add.legend = TRUE, l.legend = "obs. OC", col = c("black"), mar = c(3.1, 3.1, .5, .5), 
+                                     ylim = lim.oc, 
+                                     xlab = "Concentration exceedance percentile", 
+                                     ylab = expression(paste("OC conc. (mg l"^"-1", ")")))')
+      } else if (exi.t["ccoc"]) {
+        cp <- cp + 1
+        list.plotexpr[[cp]] <- parse(text = '.PlotDurationCurve(ExtractFreq(data = ccoc), xscale = xscale, yscale = yscale, 
+                                     add.legend = TRUE, l.legend = "sim. OC", col = c("red"), mar = c(3.1, 3.1, .5, .5), 
+                                     ylim = lim.oc, 
+                                     xlab = "Concentration exceedance percentile", 
+                                     ylab = expression(paste("OC conc. (mg l"^"-1", ")")))')
+      } else {
+        cp <- cp + 1
+        list.plotexpr[[cp]] <- parse(text = 'frame()')
+      }
+      
+      
+      ## panel 5: sim-obs regime plot for OC
+      if (exi.t["reoc"] && exi.t["ccoc"]){
+        cp <- cp + 1
+        list.plotexpr[[cp]] <- parse(text = '.PlotSimObsRegime(x = xw, sim = "ccoc", obs = "reoc", start.mon = start.mon, 
+                                     log = log.r, l.legend = c("sim. OC", "obs. OC"), mar = c(3.1, 3.1, .5, .5),
+                                     ylab = expression(paste("OC conc. (mg l"^"-1", ")")))')
+      } else if (exi.t["reoc"]){
+        cp <- cp + 1
+        list.plotexpr[[cp]] <- parse(text = '.PlotSimObsRegime(x = xw, sim = NULL, obs = "reoc", start.mon = start.mon, 
+                                     log = log.r, l.legend = c("obs. OC"), mar = c(3.1, 3.1, .5, .5),
+                                     ylab = expression(paste("OC conc. (mg l"^"-1", ")")))')
+      } else if (exi.t["ccoc"]){
+        cp <- cp + 1
+        list.plotexpr[[cp]] <- parse(text = '.PlotSimObsRegime(x = xw, sim = "ccoc", obs = NULL, start.mon = start.mon, 
+                                     log = log.r, l.legend = c("sim. OC"), mar = c(3.1, 3.1, .5, .5), ylim = lim.oc, 
+                                     ylab = expression(paste("OC conc. (mg l"^"-1", ")")))')
+      }
+    }