+++ R/functioncollection_import.R
@@ -676,10 +676,14 @@ ReadGeoData <- function(filename = "GeoData.txt", sep = "\t", encoding = c("unkn
   cnumeric <- which(toupper(cnames) %in% c("AREA", "RIVLEN"))
   
   res <- fread(filename, header = TRUE, sep = sep, colClasses = list("numeric" = cnumeric), data.table = FALSE, encoding = encoding, fill = TRUE)
-  names(res) <- toupper(names(res))
+  
+  # Save column names and make unique so that select call can remove columns with all NA if there are duplicate column names
+  resnames <- toupper(names(res))
+  names(res) <- make.unique(resnames)
   
   # remove columns with all NA
   res <- res %>% select(where(function(x){any(!is.na(x))}))
+  names(res) <- resnames # Restore column names
   
   # # NOT USED ATM, REPLACED BY colClasses ARGUMENT IN fread. LEFT FOR REFERENCE
   # # force type numeric for selected columns if they exist. Otherwise there can be problem with integer calculation in other functions..
@@ -720,14 +724,14 @@ ReadGeoData <- function(filename = "GeoData.txt", sep = "\t", encoding = c("unkn
   
   if (any(is.na(pos.m))) {
     # warn if mandatory columns are missing
-    warning(paste0("Mandatory 'HypeGeoData' column(s) '", paste(m[is.na(pos.m)], collapse = "', '"), "' missing. Imported as 'data.frame'."))
+    warning(paste0("Mandatory 'HypeGeoData' column(s) '", paste(m[is.na(pos.m)], collapse = "', '"), "' missing. Imported as 'data.frame'."), call. = FALSE)
 #    class(res) <- class(res)[-1]
     class(res) <- "data.frame"
   }
   
   if (length(pos.s) == 0) {
     # warn if there are no SLC columns
-    warning("Mandatory 'HypeGeoData' column(s) 'SLC_n' missing. Imported as 'data.frame'.")
+    warning("Mandatory 'HypeGeoData' column(s) 'SLC_n' missing. Imported as 'data.frame'.", call. = FALSE)
     if (class(res)[1] == "HypeGeoData") {
       #    class(res) <- class(res)[-1]
       class(res) <- "data.frame"
@@ -740,12 +744,12 @@ ReadGeoData <- function(filename = "GeoData.txt", sep = "\t", encoding = c("unkn
       te1 <- n.s[dn.s > 1]
       te2 <- sapply(dn.s[dn.s > 1] - 1, function(x) 1:x)
       slc.miss <- sort(unlist(sapply(1:length(te1), function(x, y, z) y[x] - z[[x]], y = te1, z = te2)))
-      warning(paste0("SLC class column(s) missing in imported file: ", paste0("SLC_", slc.miss, collapse = ", ")))
+      warning(paste0("SLC class column(s) missing in imported file: ", paste0("SLC_", slc.miss, collapse = ", ")), call. = FALSE)
     }
     if (any(dn.s == 0)) {
       # warn if there are SLC class duplicates
       warning(paste0("SLC class column duplicate(s) in imported file: ", 
-                     paste0("SLC_", n.s[dn.s == 0], " (", dn.s[dn.s == 0], ")", collapse = ", ")))
+                     paste0("SLC_", n.s[dn.s == 0], " (", dn.s[dn.s == 0], ")", collapse = ", ")), call. = FALSE)
     }
   }
   
@@ -837,6 +841,11 @@ ReadGeoData <- function(filename = "GeoData.txt", sep = "\t", encoding = c("unkn
   # Warning if any non-standard column names
   if(any(check_cols == FALSE)){
     warning("Non-standard GeoData column names found: ", paste(colnames(res)[which(check_cols == FALSE)], collapse = ", "), call. = FALSE)
+  }
+  
+  # Warning for duplicate column names
+  if(any(duplicated(colnames(res)))){
+    warning("Duplicate column names found: ", paste(colnames(res)[which(duplicated(colnames(res)))], collapse = ", "), call. = FALSE)
   }
   
   return(res)