+++ NEWS.md
@@ -1,6 +1,17 @@
 # HYPEtools NEWS
 
-## HYPEtools 1.6.4
+## HYPEtools 1.6.5
+*Enhancements and bug fixes*
+
+#### Highlights
+* Fix bug in AllUpstreamSubids() that resulted in incorrect upstream SUBIDs when MAINPART == 0 in BranchData.txt
+* Add plots for modeled organic carbon in PlotBasinOutput() and PlotBasinSummary()
+* Update ReadGeoData() to check and warn for duplicate column names
+* Add check to PlotSubbasinRouting() to ensure that "map" and "gd" have shared SUBIDs and add handling for if SUBIDs in "map" and "gd" are not identical
+* Allow for custom line colors and weights in PlotSubbasinRouting()
+* Add RunHYPE() function
+
+## HYPEtools 1.6.4 (2024-11-27)
 *Enhancements and bug fixes*
 
 #### Highlights